test: e2e setup okay
