+++ .gitignore
@@ -84,3 +84,4 @@ android/generated
 
 # React Native Nitro Modules
 nitrogen/
+example/.env
