--- detox.config.js
@@ -1,51 +1,0 @@
-module.exports = {
-  testRunner: {
-    args: {
-      $0: 'jest',
-      config: 'e2e/jest.config.js',
-    },
-    jest: {
-      setupFilesAfterEnv: ['<rootDir>/e2e/init.ts'],
-    },
-  },
-  apps: {
-    'ios.debug': {
-      type: 'ios.app',
-      binaryPath:
-        'example/ios/build/Build/Products/Debug-iphonesimulator/example.app',
-      build:
-        'xcodebuild -workspace example/ios/example.xcworkspace -scheme example -configuration Debug -sdk iphonesimulator -derivedDataPath example/ios/build',
-    },
-    'android.debug': {
-      type: 'android.apk',
-      binaryPath: 'example/android/app/build/outputs/apk/debug/app-debug.apk',
-      build:
-        'cd example/android && ./gradlew assembleDebug assembleAndroidTest -DtestBuildType=debug',
-    },
-  },
-  devices: {
-    simulator: {
-      type: 'ios.simulator',
-      device: {
-        type: 'iPhone 16 Pro',
-        os: 'iOS 18.5',
-      },
-    },
-    emulator: {
-      type: 'android.emulator',
-      device: {
-        avdName: 'Pixel_4_API_30',
-      },
-    },
-  },
-  configurations: {
-    'ios.sim.debug': {
-      device: 'simulator',
-      app: 'ios.debug',
-    },
-    'android.emu.debug': {
-      device: 'emulator',
-      app: 'android.debug',
-    },
-  },
-};
