+++ e2e/constants.ts
@@ -1,14 +1,11 @@
-// Test constants - React Native E2E testing
-// Feature IDs should match the actual feature flags configured in your Bucketeer environment
-
-export const FEATURE_TAG = 'mobile';
-export const USER_ID = 'js-e2e-user-123';
+export const FEATURE_TAG = 'javascript';
+export const USER_ID = 'bucketeer-js-user-id-1';
 
 export const FEATURE_ID_BOOLEAN = 'feature-js-e2e-boolean';
 export const FEATURE_ID_STRING = 'feature-js-e2e-string';
-export const FEATURE_ID_INT = 'js-e2e-number';
+export const FEATURE_ID_INT = 'feature-js-e2e-int';
 export const FEATURE_ID_DOUBLE = 'feature-js-e2e-double';
-export const FEATURE_ID_JSON = 'js-e2e-object';
+export const FEATURE_ID_JSON = 'feature-js-e2e-json';
 
 export const GOAL_ID = 'goal-js-e2e-1';
 export const GOAL_VALUE = 1.0;
