--- e2e/init.ts
@@ -1,2 +1,0 @@
-// Detox E2E test initialization
-// Globals are provided by Detox Jest environment
