--- e2e/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  rootDir: '..',
-  testMatch: ['<rootDir>/e2e/**/*.spec.ts'],
-  testTimeout: 120000,
-  maxWorkers: 1,
-  globalSetup: 'detox/runners/jest/globalSetup',
-  globalTeardown: 'detox/runners/jest/globalTeardown',
-  reporters: ['detox/runners/jest/reporter'],
-  testEnvironment: 'detox/runners/jest/testEnvironment',
-  verbose: true,
-  setupFilesAfterEnv: ['<rootDir>/e2e/init.ts'],
-};
