+++ example/env.template
@@ -1,0 +1,2 @@
+EXPO_PUBLIC_BKT_API_ENDPOINT=YOUR_API_ENDPOINT
+EXPO_PUBLIC_BKT_API_KEY=YOUR_API_KEY
